refactor: rename some queries
