+++ with-near-operation-file-preset/src/components/FeedItem/fragment.gql
